# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a search service server.",
        "title": "CampIn Search Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Šimen Ravnik",
            "email": "sr8905@student.uni-lj.si"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "20.13.80.52",
    "basePath": "search-service/v1",
    "paths": {
        "/health/live": {
            "get": {
                "description": "Liveness",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Liveness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.RecommendationServicePingResponse"
                            }
                        }
                    }
                }
            }
        },
        "/health/ready": {
            "get": {
                "description": "Readiness",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Readiness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Metrics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/listings": {
            "get": {
                "description": "Returns listing by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listings"
                ],
                "summary": "Listings list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Listing"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listings"
                ],
                "summary": "Listings create",
                "parameters": [
                    {
                        "description": "Listing",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.Listing"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Listing"
                            }
                        }
                    }
                }
            }
        },
        "/v1/listings/{id}": {
            "get": {
                "description": "Returns listing by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listings"
                ],
                "summary": "Listings by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Listing ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Listing"
                            }
                        }
                    }
                }
            }
        },
        "/v1/recommendations": {
            "get": {
                "description": "Returns recommendations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendations"
                ],
                "summary": "Recommendations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.LocationsResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.LocationsResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "city": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "population": {
                                        "type": "integer"
                                    },
                                    "tempC": {
                                        "type": "integer"
                                    },
                                    "weather": {
                                        "type": "string"
                                    },
                                    "weatherShort": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "error": {
                            "type": "string"
                        },
                        "traceId": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "api.RecommendationServicePingResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "db.Listing": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}